requirements.txt
flask==2.3.2
google-generativeai==0.3.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
flask-caching==2.1.0
PyPDF2==3.0.1
python-dotenv==1.0.0